Cambio para que select de programa
Updated CursoForm to use a ModelChoiceField for 'nombreprograma' and set related fields as readonly. Added a JavaScript file to auto-populate 'codigoprograma', 'versionprograma', and 'duracionprograma' based on the selected program. Introduced a new view and URL endpoint to provide program details via AJAX, and updated the template to include the new script and adjust field order.